+++ src/ducks/reducer.js
@@ -3,9 +3,9 @@ import auth from './auth';
 import programs from './programs';
 import classrooms from './classrooms';
 import assignments from './assignments';
-import mapexplorer from './mapexplorer';
+import mapexplorer from './wildcam-mapexplorer';
 import notification from './notification';
-import caesarExports from 'caesar-exports';
+import caesarExports from './caesar-exports';
 
 export default combineReducers({
   auth,
