+++ src/lib/wildcam-mapexplorer-helpers.js
